requirements.txt
absl-py==1.4.0
appdirs==1.4.4
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cloudpickle==2.2.1
cryptography==39.0.1
dill==0.3.6
grpcio==1.51.3
idna==3.4
jax==0.4.1
jaxlib==0.4.1
joblib==1.2.0
keras==2.11.0
multiprocess==0.70.14
numpy==1.24.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.4.2
pip==22.3.1
pooch==1.6.0
protobuf==3.19.4
pycparser==2.21
pyOpenSSL==23.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.1
setuptools==65.6.3
six==1.16.0
spu==0.3.1b9
threadpoolctl==3.1.0
typing_extensions==4.4.0
urllib3==1.26.14
wheel==0.38.4
